improve matrix value updating
